+++ R/tree_distance.R
@@ -39,7 +39,7 @@ GeneralizedRF <- function (splits1, splits2, nTip, PairScorer,
   
   if (reportMatching) {
     matching <- solution$matching
-    matching[matching > nSplits2] <- NA
+    matching[matching > nSplits2 | matching == 0L] <- NA
     if (nSplits1 < nSplits2) {
       matching <- matching[seq_len(nSplits1)]
     }
@@ -74,11 +74,12 @@ GeneralizedRF <- function (splits1, splits2, nTip, PairScorer,
 }
 
 .MaxValue <- function (tree1, tree2, Value) {
-  maxValue <- outer(Value(tree1), Value(tree2), '+')[, , drop = TRUE]
-  if (!inherits(tree1, 'phylo') && identical(tree1, tree2)) {
+  value1 <- Value(tree1)
+  if (is.null(tree2)) {
+    maxValue <- outer(value1, value1, '+')[, , drop = TRUE]
     maxValue[lower.tri(maxValue)]
   } else {
-    maxValue
+    outer(value1, Value(tree2), '+')[, , drop = TRUE]
   }
 }