+++ R/tree_distance_msd.R
@@ -37,7 +37,7 @@
 #' 
 #' @importFrom TreeTools LnUnrooted.int
 #' @export
-MatchingSplitDistance <- function (tree1, tree2 = tree1, normalize = FALSE,
+MatchingSplitDistance <- function (tree1, tree2 = NULL, normalize = FALSE,
                                    reportMatching = FALSE) {
   unnormalized <- CalculateTreeDistance(MatchingSplitDistanceSplits, tree1, tree2, 
                                         reportMatching)