+++ R/tree_distance_rf.R
@@ -14,6 +14,9 @@
 #' to be identical by chance alone make a smaller contribution to overall
 #' tree distance, because their similarity is less remarkable.
 #' 
+#' Rapid comparison between multiple pairs of trees employs the Day (1985)
+#' linear-time algorithm.
+#' 
 #' @inheritParams TreeDistance
 #' @param similarity Logical specifying whether to report the result as a tree
 #' similarity, rather than a difference.
@@ -32,6 +35,8 @@
 #' @references 
 #' 
 #' \insertRef{Robinson1981}{TreeDist}
+#' 
+#' \insertRef{Day1985}{TreeDist}
 #' 
 #' \insertRef{Steel2006}{TreeDist}
 #' 
@@ -59,7 +64,7 @@
 
 #' @aliases RobinsonFouldsInfo
 #' @rdname Robinson-Foulds
-InfoRobinsonFoulds <- function (tree1, tree2 = tree1, similarity = FALSE,
+InfoRobinsonFoulds <- function (tree1, tree2 = NULL, similarity = FALSE,
                                 normalize = FALSE, reportMatching = FALSE) {
   unnormalized <- CalculateTreeDistance(InfoRobinsonFouldsSplits, tree1, tree2, 
                                         reportMatching) * 2
@@ -93,14 +98,28 @@ InfoRobinsonFouldsSplits <- function (splits1, splits2,
 #' @rdname Robinson-Foulds
 #' @importFrom TreeTools NSplits
 #' @export
-RobinsonFoulds <- function (tree1, tree2 = tree1, similarity = FALSE,
+RobinsonFoulds <- function (tree1, tree2 = NULL, similarity = FALSE,
                             normalize = FALSE, reportMatching = FALSE) {
-  unnormalized <- CalculateTreeDistance(RobinsonFouldsSplits, tree1, tree2, 
-                                        reportMatching)
-  
-  if (similarity) {
-    unnormalized <- .MaxValue(tree1, tree2, NSplits) - unnormalized
+  if (is.null(tree2)) {
+    ct <- as.ClusterTable(tree1)
+    rf <- robinson_foulds_all_pairs(if(is.list(ct)) ct else list(ct))
+    if (similarity) {
+      unnormalized <- structure(rf + rf, Size = length(tree1), class = 'dist')
+    } else {
+      splits <- NSplits(tree1)
+      nSplits <- outer(splits, splits, '+')
+      unnormalized <- structure(nSplits[lower.tri(nSplits)] - rf - rf,
+                                Size = length(tree1),
+                                class = 'dist')
+    }
+  } else {
+    unnormalized <- CalculateTreeDistance(RobinsonFouldsSplits, tree1, tree2,
+                                          reportMatching)
+    if (similarity) {
+      unnormalized <- .MaxValue(tree1, tree2, NSplits) - unnormalized
+    }
   }
+  
   
   # Return:
   NormalizeInfo(unnormalized, tree1, tree2, how = normalize,
@@ -112,7 +131,7 @@ RobinsonFoulds <- function (tree1, tree2 = tree1, similarity = FALSE,
 #' @param \dots Not used.
 #' @importFrom TreeTools NSplits
 #' @export
-RobinsonFouldsMatching <- function (tree1, tree2 = tree1, similarity = FALSE,
+RobinsonFouldsMatching <- function (tree1, tree2, similarity = FALSE,
                                     normalize = FALSE, ...) {
   ret <- CalculateTreeDistance(RobinsonFouldsSplits, tree1, tree2,
                                reportMatching = TRUE)