+++ man-roxygen/tree12ListParams.R
@@ -1,2 +1,4 @@
 #' @param tree1,tree2 Trees of class `phylo`, with leaves labelled identically,
-#' or lists of such trees to undergo pairwise comparison.
+#' or lists of such trees to undergo pairwise comparison.  Where implemented,
+#' `tree2 = NULL` will compute distances between each pair of trees in the list
+#' `tree1` using a fast algorithm based on Day (1985).