+++ src/tree_distances.h
@@ -69,6 +69,7 @@ extern double
              const int16 n_different, const int16 n_a_only),
   ic_element(const int16 nkK, const int16 nk,
              const int16 nK, const int16 n),
+  ic_matching(const int16 a, const int16 b, const int16 n),
   one_overlap(const int16 a, const int16 b, const int16 n),
   one_overlap_notb(const int16 a, const int16 n_minus_b, const int16 n),
   spi_overlap(const splitbit* a_state, const splitbit* b_state, const int16 n_tips,